pycurator/requirements.txt
datasets
Flask
Flask-Cors
gunicorn
matplotlib
networkx
nltk
numpy
pandas
pydantic
pyinflect
python-dotenv
PyYAML
requests
scikit-learn
seaborn
sentence-transformers
spacy
torch
transformers
